+++ src/renderer/vulkan.rs
@@ -3,7 +3,7 @@
 //! A set of functions used to ease Vulkan resources creations. These are supposed to be internal but
 //! are exposed since they might help users create descriptors sets when using the custom textures.
 
-use crate::{Options, RendererResult};
+use crate::{Options, RenderMode, RendererResult};
 use ash::{Device, vk};
 pub(crate) use buffer::*;
 use std::{
@@ -11,9 +11,6 @@ use std::{
     mem::{self, size_of},
 };
 pub(crate) use texture::*;
-
-#[cfg(feature = "dynamic-rendering")]
-use crate::DynamicRendering;
 
 /// Return a `&[u8]` for any sized object passed in.
 pub(crate) unsafe fn any_as_u8_slice<T: Sized>(any: &T) -> &[u8] {
@@ -62,8 +59,7 @@ pub(crate) fn create_vulkan_pipeline_layout(
 pub(crate) fn create_vulkan_pipeline(
     device: &Device,
     pipeline_layout: vk::PipelineLayout,
-    #[cfg(not(feature = "dynamic-rendering"))] render_pass: vk::RenderPass,
-    #[cfg(feature = "dynamic-rendering")] dynamic_rendering: DynamicRendering,
+    render_mode: RenderMode,
     options: Options,
 ) -> RendererResult<vk::Pipeline> {
     let entry_point_name = CString::new("main").unwrap();
@@ -200,40 +196,45 @@ pub(crate) fn create_vulkan_pipeline(
         .dynamic_state(&dynamic_states_info)
         .layout(pipeline_layout);
 
-    #[cfg(not(feature = "dynamic-rendering"))]
-    let pipeline_info = pipeline_info.render_pass(render_pass);
+    let pipelines = match render_mode {
+        RenderMode::RenderPass(render_pass) => {
+            let pipeline_info = pipeline_info.render_pass(render_pass);
 
-    #[cfg(feature = "dynamic-rendering")]
-    let color_attachment_formats = [dynamic_rendering.color_attachment_format];
-    #[cfg(feature = "dynamic-rendering")]
-    let mut rendering_info = {
-        let mut rendering_info = vk::PipelineRenderingCreateInfo::default()
-            .color_attachment_formats(&color_attachment_formats);
-        if let Some(depth_attachment_format) = dynamic_rendering.depth_attachment_format {
-            rendering_info = rendering_info.depth_attachment_format(depth_attachment_format);
+            unsafe {
+                device.create_graphics_pipelines(
+                    vk::PipelineCache::null(),
+                    std::slice::from_ref(&pipeline_info),
+                    None,
+                )
+            }
         }
-        if let Some(stencil_attachment_format) = dynamic_rendering.stencil_attachment_format {
-            rendering_info = rendering_info.stencil_attachment_format(stencil_attachment_format);
-        }
-        rendering_info
-    };
-    #[cfg(feature = "dynamic-rendering")]
-    let pipeline_info = pipeline_info.push_next(&mut rendering_info);
+        RenderMode::DynamicRendering(dynamic_rendering) => {
+            let color_attachment_formats = [dynamic_rendering.color_attachment_format];
 
-    let pipeline = unsafe {
-        device
-            .create_graphics_pipelines(
-                vk::PipelineCache::null(),
-                std::slice::from_ref(&pipeline_info),
-                None,
-            )
-            .map_err(|e| e.1)?[0]
+            let mut rendering_info = vk::PipelineRenderingCreateInfo::default()
+                .color_attachment_formats(&color_attachment_formats);
+            if let Some(depth_attachment_format) = dynamic_rendering.depth_attachment_format {
+                rendering_info = rendering_info.depth_attachment_format(depth_attachment_format);
+            }
+
+            let pipeline_info = pipeline_info.push_next(&mut rendering_info);
+
+            unsafe {
+                device.create_graphics_pipelines(
+                    vk::PipelineCache::null(),
+                    std::slice::from_ref(&pipeline_info),
+                    None,
+                )
+            }
+        }
     };
 
     unsafe {
         device.destroy_shader_module(vertex_module, None);
         device.destroy_shader_module(fragment_module, None);
     }
+
+    let pipeline = pipelines.map_err(|e| e.1)?[0];
 
     Ok(pipeline)
 }