+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Breaking
 
 - Renderer is now generic over the `Allocator` trait.
+- Remove `dynamic-rendering` feature.
+- Functions that used to take `vk::RenderPass` or `DynamicRendering` now take `RenderMode` instead.
 
 ## 0.12.0